catalogs , quickserve , subscription has been done

+++ lib/features/payments/details_page.dart
@@ -10,7 +10,7 @@ class PaymentsDetailsPage extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final asyncData = ref.watch(paymentsGetProvider(id));
+    final asyncData = ref.watch(paymentsGetProvider(id as int));
     return asyncData.when(
       loading: () => _buildLoadingSkeleton(context),
       error: (e, st) => _buildErrorState(context, e.toString()),
